+++ sdmx/source/ilo.py
@@ -1,24 +1,27 @@
+import logging
+
 from . import Source as BaseSource
+
+log = logging.getLogger(__name__)
 
 
 class Source(BaseSource):
     _id = "ILO"
 
     def modify_request_args(self, kwargs):
-        """Handle two limitations of ILO's REST service.
+        """Handle limitations of ILO's REST service.
 
-        1. Service returns SDMX-ML 2.0 by default, whereas :mod:`sdmx` only
-           supports SDMX-ML 2.1. Set ``?format=generic_2_1`` query parameter.
-        2. The service does not support values 'parents', 'parentsandsiblings'
-           (the default), and 'all' for the ``references`` query parameter.
-           Override the default with ``?references=none``.
-
-           .. note:: Valid values are: none, parents, parentsandsiblings,
-              children, descendants, all, or a specific structure reference
-              such as 'codelist'.
+        1. Service returns SDMX-ML 2.0 by default, whereas :mod:`sdmx` only supports
+           SDMX-ML 2.1. Set ``?format=generic_2_1`` query parameter.
+        2. Service does not support the ``?references=…`` query parameter; discard.
         """
         super().modify_request_args(kwargs)
 
         kwargs.setdefault("params", {})
         kwargs["params"].setdefault("format", "generic_2_1")
-        kwargs["params"].setdefault("references", "none")
+
+        references = kwargs["params"].pop("references", None)
+        if references:
+            log.warning(
+                f"{self.id} does not support references={references!r}; discarded"
+            )